refactor: refactor and separate Media/ImageLogo components
